+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2022-08-28
+
+### - Added custom JSON collections using the jsonCollections property
+
+### - Enabled `/` as a collection and name divider (`carbon/add`). Unfortunately, Tailwind CSS won't pass `carbon:add` correctly.
+
 ## [0.2.0] - 2022-08-13
 
 ### - Moved icons to component layer